fix owned type bug. Fix value type structure. Refactor Expense Entity
