+++ ExpenseTrackerApi/ApiModels/ExpenseApiModel.cs
@@ -8,5 +8,5 @@ public class ExpenseApiModel
     
     public string Description { get; set; } = string.Empty;
     
-    public Money Amount { get; set; } = new Money();
+    public Money Amount { get; set; }
 }
