+++ ExpenseTrackerApi/ObjectMappingConfig/RegisterMapsterTypeConfigs.cs
@@ -10,9 +10,8 @@ public static class RegisterMapsterTypeConfigs
     public static void RegisterTypeConfiguration(this IServiceCollection services)
     {
         TypeAdapterConfig<AddExpenseModel, Expense>.NewConfig()
-            .Map(dest => dest.Amount, 
-                src => new Money(src.Amount, 
-                    new Currency("NGN", "Naira", "₦")));
+            .ConstructUsing(src => new Expense(src.Description, new Money(src.Amount, 
+                    new Currency("NGN", "Naira", "₦")), src.ExpenseDate));
         TypeAdapterConfig<ExpenseTypeApiModel, ExpenseType>.NewConfig()
             .Map(dest => dest.Id, src => src.ExpenseTypeId)
             .Map(dest => dest.Name, src => src.Name)
